+++ Tetris/src/AbstractWidget.cpp
@@ -420,7 +420,7 @@ void AbstractWidget::paintActiveBlockShadow(const Game& inGame)
 
         colIdx = block.column();
         gridPtr.reset(new Grid(block.grid()));
-        for (; rowIdx < game.rowCount(); ++rowIdx)
+        for (; static_cast<int64_t>(rowIdx) < game.rowCount(); ++rowIdx)
         {
             if (!gameState.checkPositionValid(block, rowIdx, colIdx))
             {
